Added projectiles to towers. Projectiles shoot/dmg enemies, and follow a cooldown.

+++ TDGame/src/helpz/Constants.java
@@ -22,7 +22,7 @@ public class Constants {
 			return "";
 		}
 		
-		public static float GetStartDmg(int towerType)
+		public static int GetStartDmg(int towerType)
 		{
 			switch(towerType)
 			{
